VCam post smooth: Prebake curves before applying gaussian filter. Support non subsequence takes. Add value field and adjust slider layout. PostSmoothLevel metadata to be integer to be the same as GaussianFilter KernelWidth param #jira UE-211276, UE-211697
#virtualized

[CL 32836598 by takashi kuribayashi in ue5-main branch]

+++ Engine/Plugins/VirtualProduction/VirtualCamera/Source/VirtualCamera/Public/LevelSequence/VirtualCameraClipsMetaData.h
@@ -283,5 +283,5 @@ private:
 
 	/* The asset registry tag that contains the post smooth level*/
 	UPROPERTY(EditAnywhere, Category = "Clips", BlueprintReadWrite, meta = (AllowPrivateAccess = "true"))
-	float PostSmoothLevel = 0.0f;
+	int32 PostSmoothLevel = 0;
 };
